rpi: webruntime=webosrpi4 (Change to SW decoding for all web media on all RPIs)
:Release Notes:
Change to SW decoding for all web media not only on RPi4 but all RPIs.

:Detailed Notes:
HW decoding doesn't work on RPI3 when 4.19 kernel is being used (the
same kernel as what we currently use on RPI4). Using long deprecated
and unsupported 4.4. kernel which doesn't even work on RPI3 B+ is not
an option. Disable HW decoding until it's fixed with PLAT-84754 for
both RPI4 and RPI3 to unblock the kernel upgrade.

:Testing Performed:
Video and sound works well on below apps.
- YouTube Test
- HTML5 Video
- Web Browser(Enact Browser)

:QA Notes:
N/A

:Issues Addressed:
[PLAT-84732] Change to SW decoding for all web media
[PLAT-84754] Apply media related patch to 4.19 Kernel + FKMS
[PLAT-47016] CCC: linux-raspberrypi=v4.19.66

Change-Id: I92bd87df5a0166113b4ee698b86157c1775fd8f2