webruntime=r39.3 (use default B and D instead of ${OUT_DIR} and ${OUT_DIR}/${BUILD_TYPE}/image)
:Release Notes:
This will allow the B to be correctly ignored by pseudo with:
https://git.openembedded.org/openembedded-core/commit/?id=ad8f5532ffaead9a5ad13e1034fe9e5e1b7979f4
as B gets include in PSEUDO_IGNORE_PATHS and non-standard
OUT_DIR won't and also couldn't be, when D was subdirectory
of it. Use standard layout to resolve all these issues.

:Detailed Notes:
Move all gn calls to the do_compile and leave do_install only
for installing files from OUT_DIR to the proper D.

src/webos/install/ using COPY() to install everything and COPY() is defined as:
src/build/toolchain/toolchain.gni:
  copy_command = "ln -f {{source}} {{output}} 2>/dev/null || (rm -rf {{output}} && cp -af {{source}} {{output}})"
so we would need to chown after copy, latest upstream is still using the same for COPY()
https://github.com/chromium/chromium/blob/f7438c7b9f006e1119986d52bf0f2110167dec59/build/toolchain/toolchain.gni#L97

We already fix the permissions with chmod here in recipe, so leave
the gn files to modify files in OUT_DIR and then install them to D
directly with right owner and right permissions under pseudo.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-121667] CCC: webruntime cleanup
[PLAT-121972] webruntime installs files with wrong owner and breaks pseudo

Change-Id: I818454c0bff5b1f14344a7b91819235c1f9d0289