rpm=webos4 (remove unsupported options)
:Release Notes:
--without-abi-compliance-checker --without-api-sanity-checker were valid
configure options with rpm-5, but rpm-5 was removed in 2017 with:
https://git.openembedded.org/openembedded-core/commit/?id=d0365f35303db3595688a6e5290677e02a72c54e
and only rpm-4 is being used.

:Detailed Notes:
Fixes:
ERROR: rpm-native-1_4.17.0-r0 do_configure: QA Issue: rpm-native: configure was passed unrecognised options: --without-api-sanity-checker --without-abi-compliance-checker [unknown-configure-option]

It's shown only with 4.17.0 version even when they were unrecognised for long time, because
of interesting side-effect of bdb removal in
https://github.com/rpm-software-management/rpm/commit/4290300e24c5ab17c615b6108f38438e31eeb1d0

And the interesting part is that reporting unrecognised options stops working
(enable_option_checking=no gets set in generated configure) when there is
AC_CONFIG_SUBDIRS in configure.ac (even when not actively being used as in
https://github.com/rpm-software-management/rpm/commit/4290300e24c5ab17c615b6108f38438e31eeb1d0#diff-49473dca262eeab3b4a43002adb08b4db31020d190caaad1594b47f1d5daa810L509
when not using internal bdb), see
https://git.savannah.gnu.org/cgit/autoconf.git/tree/lib/autoconf/status.m4#n1099

:Testing Performed:
Only build tested.

:QA Notes:
No change to the image.

:Issues Addressed:
[WRN-7597] CCC: Various build fixes

Change-Id: Ic098d077c6b4ace8d4a538ae024808c3b2cc16fa