wayland-protocols: remove bbappend
:Release Notes:
The patches don't apply as the wayland-procols now uses
meson instead of automake for builds.

:Detailed Notes:
The state of wl-drm patches is also a bit unclear, it wasn't merged to mesa in 2016 and
the only use of it seems to be in:
meta-webos-raspberrypi/recipes-multimedia/gstreamer/gstreamer1.0-plugins-bad/0002-waylandsink-zero-copy-using-wl-drm-protocol.patch
which wasn't ported to new gstreamer 1.18 as well and I've removed it
from meta-webos-raspberrypi/recipes-multimedia/gstreamer/gstreamer1.0-plugins-bad_1.16.%.bbappend in
hardknott with:
  commit 97b4ea1378d9d5f8c58b0e4ff781f15073ee7a66
  Author: Martin Jansa <martin.jansa@lge.com>
  Date:   Sat Oct 31 04:11:11 2020 -0700
  Subject: gstreamer*: rename bbappends to match new 1.18.0 version from oe-core

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRN-9820] Create GPVB with Yocto 3.5 Kirkstone