Revert "libtool: Update 0001-libtool-fix-clang-linking-with-fsanitize-address-or.patch"
:Release Notes:
gatesgarth and hardknott are using libtool-2.4.6 like dunfell, but don't have
the patch which adds -f*-prefix-map* parameter in the same chunk:
  3dcc84e37c libtool: Fix lto option passing for reproducible builds
causing this patch to fail to apply. In honister this change needs to be
reverted and in kirkstone this patch gets removed completely as it's
already included in upstream version.

:Detailed Notes:
This reverts commit a238074dcfdb5a8628c2b4c1f4135fd102efb67e.

:Testing Performed:
Only build tested.

:QA Notes:
None.

:Issues Addressed:
[PLAT-141293] Create GPVB with Yocto 3.4 Honister

Change-Id: I481b176585e800b5f3cae1819d6e0f403f4d7ee0