abseil=v20210324.2+git
:Release Notes:
Backport newer version which correctly packages the libraries
to avoid conflict between libabsl*.so from abseil-cpp recipe and
the libabsl*.so included in tensorflow-lite package.

:Detailed Notes:
Fixes:
 * check_data_file_clashes: Package tensorflow-lite
  wants to install file BUILD/work/qemux86_64-webos-linux/webos-image-devel/1.0-r2/rootfs/usr/lib/libabsl_bad_variant_access.so
  But that file is already provided by package  * abseil-cpp

It's still bad that tensorflow-lite doesn't use system abseil-cpp,
but for now at least avoid the conflict (abseil-cpp has these
as symlinks in abseil-cpp-dev which isn't installed by default)
and with older Yocto releases we weren't installing abseil-cpp
into the image, but now it's in grpc runtime dependencies.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[PLAT-124341] Create GPVB with Yocto 3.3 Hardknott

Change-Id: I86fb196fd9b6c2bfbce659e0f842b57b8c19d08f