meta-webos-backports-3.3: add few SPDXLICENSEMAP entries
:Release Notes:
First 3 are already included in meta-qt6/conf/layer.conf, but
that's used only by OSE builds.

:Detailed Notes:
Allow to use new license identifiers, but map them to the old
names until we upgrade to hardknott which contains the new common
licenses:
https://git.openembedded.org/openembedded-core/commit/?h=hardknott&id=5ecf139a31fa7bd813855f1235ea9f434fbcb2e0

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Build Notes:
This depends on:
http://gpro.lge.com/c/webos/build-webos/+/331296 weboslayers.py: add layer with backports from Yocto 3.3 Hardknott
http://gpro.lge.com/c/webos-pro/build-webos-pro/+/331297 weboslayers.py: add layer with backports from Yocto 3.3 Hardknott

:Issues Addressed:
[WRO-6542] CCC: Various build fixes

Change-Id: I427050c0d53d335f1e6e4ab590befff2a5f06e0e