meta-webos-backports-3.3: backport {,L}GPL-*-{only,or-later} licenses
:Release Notes:
More SPDX licenses were added to common licenses in:
https://git.openembedded.org/openembedded-core/commit/?h=hardknott&id=5ecf139a31fa7bd813855f1235ea9f434fbcb2e0
we already use some of them, backport them.

:Detailed Notes:
Fixes:
do_rootfs: The license listed GPL-3.0-or-later was not in the licenses collected for recipe lib32-gstreamer1.0-libav
do_rootfs: The license listed LGPL-3.0-or-later was not in the licenses collected for recipe lib32-gstreamer1.0-libav

:Release Notes:
Be aware of SPDXLICENSEMAP entries in meta-qt6/conf/layer.conf.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Build Notes:
This depends on:
http://gpro.lge.com/c/webos/build-webos/+/331296 weboslayers.py: add layer with backports from Yocto 3.3 Hardknott
http://gpro.lge.com/c/webos-pro/build-webos-pro/+/331297 weboslayers.py: add layer with backports from Yocto 3.3 Hardknott

:Issues Addressed:
[WRO-6542] CCC: Various build fixes

Change-Id: I427050c0d53d335f1e6e4ab590befff2a5f06e0e