meta-webos-backports-3.4: backport NPL-1.1, IJG and curl licenses
:Release Notes:
More SPDX licenses were added to common licenses in:
https://git.openembedded.org/openembedded-core/commit/?h=honister&id=e2f9092c37395f4e3ee9d0777e28c83cce6007ee
we already use NPL-1.1, IJG and curl, backport them.

:Detailed Notes:
Fixes:
do_rootfs: The license listed curl was not in the licenses collected for recipe lib32-acr2
do_rootfs: The license listed NPL-1.1 was not in the licenses collected for recipe lib32-npapi-plugins
do_rootfs: The license listed IJG was not in the licenses collected for recipe lib32-mheg

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRO-6542] CCC: Various build fixes

Change-Id: Ic100ccf87baa98b78e0b61493fa4950635c080bc