xinetd, hunspell: move bbappend to meta-webos-backports-4.2
:Release Notes:
https://lists.openembedded.org/g/openembedded-core/message/176493
was merged as:
https://git.openembedded.org/openembedded-core/commit/?id=00078ec396ff53aae71037ce7d9950afbf744454

https://lists.openembedded.org/g/openembedded-devel/message/100861
was merged as:
https://git.openembedded.org/meta-openembedded/commit/?id=81d607b2676c887447130a626f1ad1268bd5d519

:Detailed Notes:
It was merged in Yocto 4.2 Mickledore, so we can move it from meta-webos
to meta-webos-backports-4.2 so that I don't need to remove it once
we upgrade to Mickledore.

Fixes issue when the fixes are applied from both places:
http://gecko.lge.com/Errors/Details/529522
ERROR: QA Issue: hunspell-ispell is listed in PACKAGES multiple times, this leads to packaging errors. [packages-list]

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-8053] CCC: Various build fixes

Change-Id: I365cdfdf829ff14b4312f688f1894eded1a8fb54
Reviewed-on: http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/344648
Reviewed-by: Multi Layer Checker <multi_layer@lge.com>
Reviewed-by: Hyunjae Shin <hyunjae92.shin@lgepartner.com>
Tested-by: Hyunjae Shin <hyunjae92.shin@lgepartner.com>