com.webos.service.unifiedsearch=r3 (add partial support for procps-4)
:Release Notes:
Allow to build with procps-4, but ther is still a TODO for component
owner to resolve.

:Detailed Notes:
Fixes build with procps-4:
http://gecko.lge.com:8000/Errors/Details/551376

-- Checking for module 'libprocps'
--   No package 'libprocps' found

there might be many more changes needed to adapt to procps 4 API, see:
https://gitlab.com/procps-ng/procps/-/issues/239

procps upgrade was now merged in oe-core together with igt-gpu-tools adaptation in:
https://git.openembedded.org/openembedded-core/commit/?id=631eba02d64f1a42514e0ae4361bbecc5cce5fa7

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-10874] Prepare for procps 4 API

Change-Id: I8a1229acdf7b95d69bb67f93b243365094762fb0