ose: *.patch: Move Upstream-Status bellow ---
:Release Notes:
There is no agreement even in upstream whether these belong to regular
commit message or bellog ---, but after seeing many changes merged in
gerrit with Upstream-Status in commit message I'm firmly in the ---
camp now.

:Detailed Notes:
The disadvanatage is that when manually rebasing patches in git or with
devtool, the existing Upstream-Status lines are lost and you need to
make sure not to let them deleted after "git am" + "git format-patch"
cycle, but you should be reviewing rebased patches anyway so hopefully
people will notice.

With Upstream-Status in the regular commit message on the other hand
I've seen many reviews on gerrit (which were already uploaded by me
explicitly without Upstream-Status, the developer refreshed them with
"git am" and push, without reviewing or understanding that the only
change of added Upstream-Status is unwanted one and should be removed
again before merging the review in gerrit (which then gets another
link to the review in "Reviewed-On" footer note from gerrit).

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-310] Apply .patch files for hot-fix to corresponding repositories
[WRP-11864] CCC: Various build fixes and .patch files cleanup
[WRP-310] Apply .patch files for hot-fix to corresponding repositories
[WRP-11630] Add Upstream-Status to existing .patch files in meta-lg-webos