uninative: call patchelf-uninative only when needed
:Release Notes:
mke2fs.real, mkfs.ext2.real, mkfs.ext3.real, mkfs.ext4.real are indentical
binary with multiple hardlinks and we end calling patchelf-uninative 4
times even when the interpreter is already set correctly from the build

:Detailed Notes:
To avoid corrupted binaries created on 18.04 ubuntu avoid calling
patchelf-uninative multiple times and in this case don't call it at all.

It might be related to:
https://github.com/NixOS/patchelf/issues/492
or
https://github.com/NixOS/patchelf/issues/446
but the later was already included in patchelf-0.17.2 used in uninative-3.9

This was submitted to upstream in:
https://lists.openembedded.org/g/openembedded-core/message/183314
but wasn't merged yet (so it cannot be in meta-webos-backports-* layer)
and it might take a while until it's backported to kirkstone.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-19053] CCC: Various build fixes
[WRP-17893] mkfs.ext4 segfaults with uninative 3.10 and newer
[WRP-6209] Update jenkins slaves to use Ubuntu 20.04 or 22.04

Change-Id: Ied1e0965423c660bca375c1e8deac7500014cc03