ose: webos_base.bbclass: Add packageconfig to webos-bom.json
:Release Notes:
Packageconfig information is useful to analyze webos image.

:Detailed Notes:
PACKAGECONFIG is a very important variable in determining which features
to enable/disable and which options to put in the do_config task before
code compilation.

Since packageconfig could not be seen in the existing webos-bom.json,
there was a risk of misunderstanding that it was the same environment
because the version and revision all matched even though it was built
with a different configuration.

PACKAGECONFIG in the webos-bom will be used to compare the different
images.

:Testing Performed:
Build test done.

:QA Notes:
N/A

:Issues Addressed:
[WRP-26053] CCC: Various build fixes

Change-Id: I3f43ce7159086a03c16bdf5a0208a9f1dc00c12c