tcl, mongodb, opengl-es-cts, libcyusbserial, tcpreplay: move to meta-webos-backports
:Release Notes:
These changes are now merged upstream, add a link to the commit there
instead of my post on ML and move them to corresponding layer in
meta-webos-backports.

:Detailed Notes:
tcl is already backported to mickledore. For other components I've just
requested backport to mickledore and added link to the ML there.
There is also libiio_%.bbappend and llvm-project-source.bbappend in
meta-webos with backports, but those weren't merged upstream yet, so
cannot be moved to meta-webos-backports yet.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-20118] Bitbake world issues 2023-07
[WRP-26053] CCC: Various build fixes

Change-Id: I568f892c90a93668e2775bfaab32c85873eebac4