webos_oss_pkg_info.bbclass: fix path to license.manifest
:Release Notes:
SSTATE_PKGARCH subdirectory was added in:
https://git.openembedded.org/openembedded-core/commit/?id=1a4ab9fc26659507e678e87312b514e8ea515673

:Detailed Notes:
And as the oe-core commit says:
For anyone finding this commit, I'd question how much people should be relying on
this code for tooling and suggest the SPDX manifests should be the preferred data
format going forward anyway.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRP-13301] Create GPVB with Yocto 4.3 Nanbield

Change-Id: I198df79c32bb9af32fdd9f79125ba8118cdaba3a