webos: webos.inc: ERROR_QA: add patch-status and virtual-slash
:Release Notes:
virtual-slash was added in:
https://git.openembedded.org/openembedded-core/commit/meta/classes-global/insane.bbclass?h=scarthgap&id=968ffdb9fee5017eecce36ce878ea604c869ce95
patch-status is enabled for patches in oe-core for very long time
and I was using it for many other layers as well, so most layers
should be in good shape, I've sent patch to enable it for all layers:
https://lists.openembedded.org/g/openembedded-core/message/197098

:Detailed Notes:
Only 4 entries are left in WARN_QA:

buildpaths - as we have many issues in our own layers as well :(
32bit-time - as we have many issues in our own layers as well :(
missing-metadata - checks SUMMARY and HOMEPAGE for oe-core recipes:
  https://git.openembedded.org/openembedded-core/commit/meta/classes-global/insane.bbclass?h=scarthgap&id=4144c2f43da39336b03cfd612cbe1694cbf8c7bd
missing-maintainer - checks RECIPE_MAINTAINER for oe-core recipes:
  https://git.openembedded.org/openembedded-core/commit/meta/classes-global/insane.bbclass?h=scarthgap&id=675dff0e37666c1cffa10a83f6f1f67d5fadd204

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRQ-504] Create GPVB with Yocto 5.0 Scarthgap

Change-Id: Iafce7894554d173dcefdedcd319d10e323d3f14d