webos: webos-backports-5.2: add backports layer for 5.2 Walnascar
:Release Notes:
Add explicit dependency in meta-webos to make sure that
everybody executes mcf --command configure before the layer
is used will be added later, when the layer is enabled in
weboslayers.py and isn't empty.

:Detailed Notes:
This way we don't need to add the layer and update weboslayers.py at the
same time (in the same CCC).

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRR-8591] CCC: webos-backports-5.2: add backports layer for 5.2 Walnascar

Change-Id: I6f385bd5a624607f4650f13806b1ab85eb339ba4