webos: luna-sysservice=r13 (use EXTRA_INHERIT)
:Release Notes:
Revert "webos: luna-sysservice=r12"
This reverts commit d38ef4d8ded6a405de98dd04a766fbd75514bf28.

Migrate this to inherit_defer after upgrading to scarthgap with:
https://git.openembedded.org/openembedded-core/commit/?h=scarthgap&id=451363438d38bd4552d5bcec4a92332f5819a5d4
https://git.openembedded.org/bitbake/commit/?h=2.8&id=5c2e840eafeba1f0f754c226b87bfb674f7bea29
now it needs to be set in DISTRO config.

:Detailed Notes:
luna-sysservice has qtbase in DEPENDS and doesn't build with
webos_cmake with or without qtbase:

http://gecko.lge.com:8000/Errors/Details/851894
http://gecko.lge.com:8000/Errors/Details/851892

So this will need to be fixed differently. Add PACKAGECONFIG
to control the inherit as well as qtbase dependency, leave the
qt enabled by default (as this version doesn't build without it
anyway) and other layers with different luna-sysservce can
disable it by setting PACKAGECONFIG to empty.

Similar to:
http://gpro.lge.com/c/test/meta-lg-webos/+/72036

:Testing Performed:
No effect on images.

:QA Notes:
N/A

:Issues Addressed:
[WRQ-21716] CCC: Various build fixes
[WRQ-21679] Provide a mechanism to allow installation of qt components

Change-Id: Ida42690d34c9a62de7936e4df20e61bcff3e0599
Reviewed-on: http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/409051
Reviewed-by: Ghane Simit <simit.ghane@lge.com>
Reviewed-by: Abdul Khader <abdul.khader@lge.com>
Reviewed-by: Hieu Van Nguyen <hieu2.nguyen@lge.com>
Reviewed-by: Jaekyu Lee <jaekyu.lee@lge.com>
Reviewed-by: Yangwoo Byun <yw.byun@lge.com>