webos: webos.inc: backport UNPACKDIR variable from styhead
:Release Notes:
This only backports the variable which is equal to WORKDIR as it was in:
https://git.openembedded.org/openembedded-core/commit/?id=e022d62ba917790af2121da57646271ef17c03fa
it doesn't even change unpack task in base.bbclass to use it not changes
the value from WORKDIR to something else as implemented in:
https://git.openembedded.org/openembedded-core/commit/?id=b84eec5c4cbf4b39d6712800dd0d2fe5337721cb
because that would require many changes in all layers.

:Detailed Notes:
This really just allows us to use UNPACKDIR variable when referring to
files "unpacked" from SRC_URI instead of using WORKDIR variable name.

This way when rebasing the changes for styhead compatibility we'll have
much fewer conflicts (as WORKDIR/UNPACKDIR is used in a lot of recipes)
and developers can learn to use this variable in advance (before it's
mandatory with styhead where WORKDIR != UNPACKDIR).

We'll need to rebase only much smaller diff for e.g. S = WORKDIR fixes
as in:
https://git.openembedded.org/openembedded-core/commit/?id=d9328e3b0b062f0621de3d114584b44f384a1c02
because that would be more difficult (or confusing) to backport.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRQ-22934] CCC: webos.inc: backport UNPACKDIR variable from styhead
[WRQ-17959] Create GPVB with Yocto 5.1 Styhead

Change-Id: I48faef775c6f3a63fdca087b4251cc671492bab5
Reviewed-on: http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/409306
Reviewed-by: Ghane Simit <simit.ghane@lge.com>
Reviewed-by: Abdul Khader <abdul.khader@lge.com>
Reviewed-by: Hieu Van Nguyen <hieu2.nguyen@lge.com>
Reviewed-by: Jaekyu Lee <jaekyu.lee@lge.com>
Reviewed-by: Yangwoo Byun <yw.byun@lge.com>