webos: webos.inc: update ERROR_QA
:Release Notes:
Most of DEFAULT_WARN_QA were moved to ERROR_QA in upstream with:
https://git.openembedded.org/openembedded-core/commit/?id=b79b191cc43a45dde2adb61ea349b426cb2461d1
so we no longer append whole DEFAULT_WARN_QA to ERROR_QA, just
native-last (which was already there).

:Detailed Notes:
We cannot move 32bit-time to ERROR_QA, because there are still many
issues in various components as shown in:
[WRQ-12953] Fix build with 64-bit time_t

and we cannot keep buildpaths in ERROR_QA (which was enabled in
oe-core with this commit) because couple recipes from meta-oe (and
even libgfrontend from oe-core) still trigger this issue (should be
fixed upstream) and other layers like meta-qt6 with:
https://bugreports.qt.io/browse/QTBUG-105913?focusedId=781605&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-781605

buildpaths will be moved from ERROR_QA to WARN_QA in follow-up
commit only for recipes which currently trigger these issues.

The list of these issues (as warnings) is in:
[WRP-10883] buildpath QA issues
208 warnings without "qt" and 854 warnings with "qt"

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRQ-17959] Create GPVB with Yocto 5.1 Styhead
[WRQ-12953] Fix build with 64-bit time_t
[WRP-10882] Resolve buildpaths QA warnings for more reproducible builds
[WRP-10883] buildpath QA issues
[WRQ-14472] bluetooth: Resolve buildpaths QA warnings
[WRP-10884] libpbnjson: Resolve buildpaths QA warnings
[WRP-31951] webos-connman-adapter: Resolve buildpaths QA warnings

Change-Id: Ib8e831f7a97ea61def77216c6f257334c31119a0