webos: nodejs-module-nan-native: remove unused recipe
:Release Notes:
It was used only by node-segfault-handler which was removed in:
http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/375406
so we don't need this recipe anymore.

:Detailed Notes:
Surprisingly it built with nodejs-22 even when it's this old version
in usocket which was using 2.18.0 I had to update to 2.22.0 version first
with:
https://github.com/nodejs/nan/commit/6bd62c9a0004339d5d1e18a945c84929d0f6b808
to fix build with newer nodejs-22 from meta-oe.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRR-5443] Create GPVB with Yocto 5.2 Walnascar

Change-Id: If4c7b59a14168a3cacc837c8e8319bace4067923