webos: libedgetpu=r10 (makefile_build/Makefile: drop -fuse-ld=gold)
:Release Notes:
gold is marked as deprecated in GNU Binutils 2.44 and will be removed in future release:
https://lists.gnu.org/archive/html/info-gnu/2025-02/msg00001.html
https://lists.openembedded.org/g/openembedded-core/message/210627
https://lists.openembedded.org/g/openembedded-core/message/210940

We've already switched from gold to bfd in:
http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/362582 webos: webos-features-core-boot.inc: switch from gold linker to default bfd

so we shouldn't be using it here as well.

:Detailed Notes:
Support for gold was removed from oe-core in:
https://git.openembedded.org/openembedded-core/commit/?id=a4addb9ab63011e7c604fc5daff95559e7d214e7

and this is now failing with:
http://gecko.lge.com:8000/Errors/Details/1109731

Building libedgetpu.so
collect2: fatal error: cannot find 'ld'
compilation terminated.
make: *** [Makefile:201: libedgetpu] Error 1

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[WRR-19944] [NANO] CCC: Yocto 5.0 Scarthgap
[WRR-2613] Switch from gold linker to lld
[WRR-5443] Create GPVB with Yocto 5.2 Walnascar

Change-Id: I6cd39e95b88d79871426224269b02b3e303a6e9c