webos: layer.conf: add dependency on webos-backports-5.2 layer
:Release Notes:
It's not strictly needed for parsing meta-webos layer and currently has:
libabigail, jansson, ot-br-posix
recipes, but it's better to record it as well as all our builds are
always including it (so nobody would notice build failure once
meta-webos starts really depending on it).

:Detailed Notes:
e.g. without the lapack backport it would fail to build with buildpaths
issue, but lapack fix was backported to scarthgap since then and dropped
from webos-backports.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[GBW-2742] [NANO] CCC: Yocto 5.0 Scarthgap

Change-Id: I79b24ccbfc9d33dfdc09c9a2f677f650a01d73b8