webos: mongodb=webos1 (lower buildpaths from ERROR_QA to WARN_QA)
:Release Notes:
This one is new because the SKIP_RECIPE was removed in scarthgap:
https://lists.openembedded.org/g/openembedded-devel/message/115780
but it's still skipped in styhead and newer because of unresolved
review comments:
https://lists.openembedded.org/g/openembedded-devel/message/116508

:Detailed Notes:
Works around:
http://gecko.lge.com:8000/Errors/Details/1155237
ERROR: QA Issue: File /usr/bin/mongos in package mongodb contains reference to TMPDIR
File /usr/bin/mongod in package mongodb contains reference to TMPDIR [buildpaths]

Until mongodb is fixed in master and the buildpaths fix could be
backported to scarthgap.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[GBW-2742] [NANO] CCC: Yocto 5.0 Scarthgap
[WRR-2120] Create GPVB with Yocto 5.1 Styhead
[WRP-10882] Resolve buildpaths QA warnings for more reproducible builds
[WRP-10883] buildpath QA issues

Change-Id: I80952e158d97ff6bd6f80afe5fd6d0ae143f2d98