webos: llvm: drop the bbappend
:Release Notes
It was added in 2017
http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/183318 llvm=webos4

New llvm-5.0 autodetects libedit, it's not an issue with newer Yocto
which uses RSS, but here we need to add PACKAGECONFIG to explicitly
disable it.

Fixes:
ERROR: QA Issue: libllvm5.0-llvm-5.0 rdepends on libedit, but it isn't
  a build dependency, missing libedit in DEPENDS or PACKAGECONFIG?
  [build-deps]
as detected in:
http://caprica.lgsvl.com:8080/Errors/Details/873319

:Detailed Notes:
The libedit dependency was added in 2019:
https://git.openembedded.org/openembedded-core/commit/?h=dunfell&id=3c08b638348abd543fc92baf56c28ca16ae6aac6
we don't really need to disable it now, so drop the PACKAGECONFIG.

Helps when llvm recipes are renamed to clang with:
https://lists.openembedded.org/g/openembedded-core/message/215499

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[GBW-3659] [NANO] CCC: Yocto 5.0 Scarthgap

Change-Id: I3d9a38a1c2c28d6e8a23f7948eecf69b9b507e78
Reviewed-on: http://kairos-git.lge.com/c/webos-pro/meta-lg-webos/+/471819
Reviewed-by: Ghane Simit <simit.ghane@lge.com>
Reviewed-by: Abdul Khader <abdul.khader@lge.com>
Reviewed-by: Hieu Van Nguyen <hieu2.nguyen@lge.com>
Reviewed-by: Youngseok Jeong <youngseok1.jeong@lge.com>